fileIO and openSSL now have better error handling / throwing.

+++ fileIO.py
@@ -31,4 +31,5 @@ def readFile(path):
 
 # Removes a file
 def removeFile(path):
-	os.remove(path)
+	if fileExists(path):
+		os.remove(path)
